+++ src/dsp_tools/commands/excel2json/lists_compliance_checks.py
@@ -1,12 +1,19 @@
 from __future__ import annotations
 
 import warnings
+from typing import Any
+from typing import cast
 
 import pandas as pd
 import regex
 
-from dsp_tools.commands.excel2json.models.input_error import ListExcelComplianceProblem
+from dsp_tools.commands.excel2json.models.input_error import ListExcelProblem
 from dsp_tools.commands.excel2json.models.input_error import ListSheetComplianceProblem
+from dsp_tools.commands.excel2json.models.input_error import ListSheetContentProblem
+from dsp_tools.commands.excel2json.models.input_error import MissingNodeTranslationProblem
+from dsp_tools.commands.excel2json.models.input_error import MissingTranslationsSheetProblem
+from dsp_tools.commands.excel2json.models.input_error import Problem
+from dsp_tools.commands.excel2json.new_lists import _get_all_languages_for_columns
 from dsp_tools.commands.excel2json.new_lists import _get_hierarchy_nums
 from dsp_tools.commands.excel2json.new_lists import _get_lang_string_from_column_name
 from dsp_tools.models.custom_warnings import DspToolsUserWarning
@@ -14,7 +21,7 @@ from dsp_tools.models.custom_warnings import DspToolsUserWarning
 
 def make_all_formal_excel_compliance_checks(
     excel_dfs: dict[str, dict[str, pd.DataFrame]],
-) -> list[ListExcelComplianceProblem]:
+) -> list[ListExcelProblem]:
     """
     This function checks if the excel files are compliant with the expected format.
 
@@ -32,16 +39,14 @@ def make_all_formal_excel_compliance_checks(
     ]
 
 
-def _make_formal_excel_compliance_check(
-    excel_dfs: dict[str, pd.DataFrame], excel_name: str
-) -> ListExcelComplianceProblem | None:
+def _make_formal_excel_compliance_check(excel_dfs: dict[str, pd.DataFrame], excel_name: str) -> ListExcelProblem | None:
     problems = [p for sheet_name, df in excel_dfs.items() if (p := _df_shape_compliance(df, sheet_name)) is not None]
     if problems:
-        return ListExcelComplianceProblem(excel_name, problems)
+        return ListExcelProblem(excel_name, problems)
     return None
 
 
-def _df_shape_compliance(df: pd.DataFrame, sheet_name: str) -> ListSheetComplianceProblem | None:
+def _df_shape_compliance(df: pd.DataFrame, sheet_name: str) -> Problem | None:
     problems = {}
     problems.update(_check_minimum_rows(df))
     problems.update(_check_min_num_col_present(df.columns))
@@ -104,3 +109,78 @@ def _check_all_expected_translations_present(cols: pd.Index[str]) -> dict[str, s
             f"{', '.join(missing_cols)}"
         }
     return {}
+
+
+def make_all_list_content_compliance_checks(
+    excel_dfs: dict[str, dict[str, pd.DataFrame]],
+) -> list[ListExcelProblem]:
+    """
+    This function checks if the content of the excel files is compliant with the expected format.
+
+    Args:
+        excel_dfs: dictionary with the excel file name as key
+                    and a dictionary with the sheet name as key and the dataframe.
+
+    Returns:
+        A list with the problems, or an empty list if there are no problems
+    """
+    return [
+        res
+        for filename in excel_dfs
+        if (res := _make_content_excel_compliance_check(excel_dfs[filename], filename)) is not None
+    ]
+
+
+def _make_content_excel_compliance_check(
+    excel_dfs: dict[str, pd.DataFrame], excel_name: str
+) -> ListExcelProblem | None:
+    problems = [p for sheet_name, df in excel_dfs.items() if (p := _df_content_compliance(df, sheet_name)) is not None]
+    if problems:
+        all_problems = [cast(Problem, p) for p in problems]
+        return ListExcelProblem(excel_name, all_problems)
+    return None
+
+
+def _df_content_compliance(df: pd.DataFrame, sheet_name: str) -> ListSheetContentProblem | None:
+    problems = []
+    if missing_translations := _check_all_nodes_translated_into_all_languages(df, sheet_name):
+        problems.append(missing_translations)
+    if problems:
+        all_problems = [cast(Problem, p) for p in problems]
+        return ListSheetContentProblem(sheet_name, all_problems)
+    return None
+
+
+def _check_all_nodes_translated_into_all_languages(
+    df: pd.DataFrame, sheet_name: str
+) -> MissingTranslationsSheetProblem | None:
+    col_endings = [str(num) for num in _get_hierarchy_nums(df.columns)]
+    col_endings.append("list")
+    languages = _get_all_languages_for_columns(df.columns)
+    all_cols = _make_columns(col_endings, languages)
+    problems = []
+    for column_group in all_cols:
+        problems.extend(_check_one_hierarchy(column_group, df))
+    if problems:
+        return MissingTranslationsSheetProblem(sheet_name, problems)
+    return None
+
+
+def _check_one_hierarchy(columns: list[str], df: pd.DataFrame) -> list[MissingNodeTranslationProblem]:
+    problems = []
+    for _, row in df.iterrows():
+        if problem := _check_one_node_for_translations(row, columns):
+            problems.append(problem)
+    return problems
+
+
+def _make_columns(nums: list[str], languages: set[str]) -> list[list[str]]:
+    return [[f"{lang}_{num}" for lang in languages] for num in nums]
+
+
+def _check_one_node_for_translations(row: pd.Series[Any], columns: list[str]) -> MissingNodeTranslationProblem | None:
+    missing = row[columns].isna()
+    if missing.any() and not missing.all():
+        missing_cols = [str(index) for index, is_missing in missing.items() if is_missing]
+        return MissingNodeTranslationProblem(missing_cols, row["row_number"])
+    return None
